+++ CHANGELOG.md
@@ -2,6 +2,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Provide functions for concatenation `concatRow`, `concatLines`, and
+  `concatGrid`. (#10)
+
 
 ## [0.9.0.0] - 2020-04-10